Add autojump to urgent to rc.lua
I added this so that when I get a link in telegram or edit a text field
in firefox with "It's All Text!", awesome automatically switches to the
tag where the link opened (in the case of the former) or the IAT buffer
is (in the case of the latter). Saves me a few keystrokes!