Add awesome custom theme
The wibar has been restyled based on the multicolor theme by @copycat-killer.
This provided an opportunity to move a lot of the styling configuration,
including the creation of the wibar and widgets, to a custom theme.
Additionally, my homegrown sysmonitor and update_notifications widgets have been
replaced by vicious widgets; accordingly, there is now a dependency on the
vicious library, which has been included as a submodule. Using the vicious pkg
widget somehow solves the async issues I have been having with
update_notifications for so long. Using a custom theme also sets the wallpaper
automatically, so `sh ~/.fehbg` is no longer required. Lastly, descriptions have
been added to several keybindings so that they show up in the new hotkeys help menu.