Add firefox keybindings to change the indentation of tabs
This fixes a really annoying issue I've had for a while where certain operations
with the tree-style tabs extension create tabs as children of the last focused
tab (so they appear indented in the tab tree), but there's no way to make the
new tab not a child (i.e. to promote the tab to the top level). Luckily such a
function exists in the tree-style tabs API, and I bound it to a key, along with
the demotion function.