Swap org-agenda for mobile notes in custom org view
SPC o o used to open stuffff.org on the left and my custom agenda on the right,
but I've changed it so that current.org is on the left (this is effectively the
same because current.org is my main org file now) and the notes synced from my
phone are on the right. This greatly reduces friction when doing my daily
review (The notes used to not be synced from my phone to my laptop, so I would
have to prop up my phone next to my laptop and transcribe the notes manually).

I've also changed the spacemacs keybinding SPC b s to open this same notes file
instead of the scratch buffer, since I've had problems in the past where I've
written down a bunch of notes in the scratch buffer and then restarted emacs and
lost them.