GTD revamp
I've done an extensive revamp of my org-mode settings in order to reduce
friction when using org-mode to manage my life. In the process, my use of
org-mode has come to more closely resemble David Allen's GTD method. One of the
biggest changes I made was completely restructuring my org files, including the
names of the files, the directory structure, and the structure of the content
within the files. See the changelog for more information on the motivations and
benefits of this restructuring.

Incomplete list of changes:
    - org-agenda-files has been updated to reflect the recent restructuring of my
      home directory and the restructuring of my org files
    - the SCHED and HW todo keywords have been eliminated, and the HABIT ->
      COMPLETE todo cycle has been added
    - habit tracking has been enabled
    - targets have been added to org-refile-paths to make refiling viable
    - many capture templates have been added
    - org-id-locations-file has been moved inside my org directory so it is
      synced to all my devices; this makes internal links by id not break when I
      switch devices
    - the option org-special-ctrl-a/e has been enabled, making it easier to
      navigate org files
    - several functions have been added in an attempt to create a simple and
      easy-to-manage project definition, but this is still WIP
    - the custom commands for the org agenda have been changed, but this is
      still a work in progress