Change face of unsaved buffers in helm-mini
This was apparently changed when I updated spacemacs. The helm-mini face for
unsaved buffers became a light grey, the same as comments in prog-mode buffers.
Needless to say I felt that it was best to go back to the old bright orange
face. If a buffer is unsaved, I want to know about it!