Add zl and zzl fasd aliases in zsh
zl will use fasd to cd somewhere and will also list the contents of that
directory. zzl does the same thing but it does it interactively (like fasd's `s`
alias). Note also that the mkdcd function has been renamed to mkcd. I hope this
makes it easier to remember so that I actually use it.