Move awesome wm widgets config to rc.lua; add custom textclock widget
The textclock widget shows decimal time by default and shows standard time while
Win-Alt-c is held down. I have also changed the date format (it's YYYY-MM-DD
now), and opted to display the day of the week using Japanese kanji.

While writing the widget I realized that configuration for all widgets would be
easier if moved to rc.lua; you're not supposed to put that stuff in the theme
file.