Switch to built-in awesome wm calendar widget; clean up files
With the latest release of awesome (4.2) there's a calendar widget provided. It
appears to have all the same functionality as the random widget I got online, so
I made the switch. I therefore removed cal.lua, as well as a couple other files
I realized I wasn't using anymore. The caffeine widget has been moved to its own
directory for consistency with the directory layouts of alttab and vicious.