Custom textclock: Fix decimal time conversion error and add colors
The error was caused by incorrectly parsing the value of os.clock in order to
add millisecond precision to the value to convert to decimal time. The widget
only updates every 3 seconds so I just got rid of the millisecond precision. If
I ever want to see decimal seconds in the widget (unlikely), I'll add it back.

The day of the week kanji is now orange and the time is pure white (to make it
stand out slightly from the date, which is grey).