Fix broken zsh powerline by porting to offical arch package
The official arch package for powerline didn't exist when I first installed
it (AFAIK), so I was using the pip solution for a long time. That way broke with
the release of python 3.7, so I had the opportunity to move to the
offically-supported installation method in the course of fixing it. Powerline
should be a lot easier to install and manage now, because it should be upgraded
along with python now, instead of me having to re-install it every time python
jumps a minor version.