Change the way that applications are spawned at startup
Screens and tags for autostarted applications are now specified directly in the
autostart command, instead of indirectly via a rule. This means that firefox
doesn't have to always be in the web tag, for example, and that IRC will spawn
correctly in its own tag. The improved spawn_once function is due to
github.com/alfunx.