New config system: fix dotbot install script and remove unused scripts
The dotbot install script assumes it is located in your repo root, but in my
case I put it inside the scripts/ directory, so I had to change the line that
sets the BASEDIR so that the script could find the files it needs.

I also removed the do_diff and configure_dotfiles scripts, since I won't be
needing these anymore.