Spacemacs: Rearrange org-mode config to match layout of manual
I have completely overhauled the organizational scheme of my org-mode config to
match that of the org-mode manual. I think this has improved the organizational
clarify of the config, and serendipitously, it's also resulted in fewer
top-level sections (which I guess is part of the increased organizational
clarity).

Important note: To make the git history nice, there are no code changes in this
commit. It's just rearranging the existing code.